fix: deterministic sort, non-tree nav label, integer order
- Sort projects and code collections by id for stable sidebar ordering
- Relabel the nav as 'Files' since it isn't an ARIA tree widget yet
- Enforce integer for pages.order

Copilot-Session: 60c01369-c856-4c09-969f-b9a693a216c5

+++ src/content.config.ts
@@ -30,7 +30,7 @@ const pages = defineCollection({
   schema: z.object({
     title: z.string(),
     // Ordering within the file listing / navigation.
-    order: z.number().default(0),
+    order: z.number().int().default(0),
     description: z.string().optional(),
   }),
 });
